Implement a deprecation message handler for the unit test library
Based on Symfony's PhpUnit bridge, this handler will catch E_USER_DEPRECATED and E_DEPRECATED messages triggered while running the unit test suite and display them at the end of the test output.  This removes the need to silence deprecation messages in the unit test bootstrap and gives contributors an idea of deprecated code uses either in the core PHP platform or libraries.

Implementing this exposed a call to a non-static method in a static manner which only affects the unit test suite, this is also corrected.