Simplification for the release process
Simplification for the release process

# Executive summary
This PR simplifies the process of packing a release. We always comment out the installer check when we are in development. This needs so far a manuell change of two files. When we are doing a stable release we also set the variable DEV_STATUS = Stable. So we have the information at one place, but we don’t use it.

# Backwards compatibility
Full b/c, no problems are expected

# Translation impact
none

# Testing instructions
Clone the joomla-cms repro and apply the patch. Do an installation. You should be able to login into the backend without removing the installation folder. Now go to /libraries/cms/version/version.php and set the DEV_STATUS variable to Stable. If you now try to access the site the system tells you that you have to remove the installation folder. You can also delete the configuration.php in the root of the Joomla install and do an install, at the end it forces you to remove the installation folder.