Allows visible text fields within JSortableList
JSortableList is altering all text fields found in the table.. which is
unnecessary and quite restrictive.
Of course it would be better to use the following selector
".text-area-order" but as there are some locations (ie. com_menus,
com_categories, ..) without the class, and maybe others in 3rd party
components, the easiet way is maybe to search for: [type=text]:hidden
At least, this allows us to have "visible" text fields without
alteration of their value.