Update uri.php
When SSL is offloaded (or SSL termination) by a load balancer, the redirect to force HTTPS is handled by the web server by using the X-Forwarded-proto X header. When using Amazon Elastic Load Balancers, you can terminate SSL and direct all traffic to non secure port on the server. If you also use $force_ssl on Joomla it causes a redirect loop. This code will allow Joomla to detect the X header if it is set by a load balancer.

https://en.wikipedia.org/wiki/List_of_HTTP_header_fields
X-Forwarded-Proto: https 
A de facto standard for identifying the originating protocol of an HTTP request, 
since a reverse proxy (load balancer) may communicate with a web server using HTTP 
 even if the request to the reverse proxy is HTTPS.