Random article order in blogs
This PR answers a fairly common request on the forums to add a random
ordering option for articles in the blog category view.

To test create a new category and in the blog options go to article
order and you will see a new option “Random Order”

Select this option and then check the article order on the front end.
Refresh the page and the article order will have changed