Using of an absolute image path
I think, it would be reasonable to start an image path with a slash. When I developed an extension, I found that the links which had been added in the frontend declaration from the media manager were interpreted by JavaScript as relative. As a result, I got 404 errors, because the paths to the files were incorrect.