Fixed issue with countModules also counting empty modules
Templates use the ```$this->countModules('myposition')``` to determine whether or not to show a position (and surrounding html).
However, when a module is empty (no content), the ```countModules``` is not affected.
This can result in positions being shown with no modules in it.

This PR fixes the issue by only counting the modules that do not have an empty content.

#### Testing Instructions
Easiest way to test with an empty module is to use Advanced Module Manager, as that will make the content truly empty even if the module has an empty ```<p></p>``` tag.
So just make a Custom HTML Module and leave the content area empty. Assign it to a module position (that has no other modules assigned to it). Like manually entering ```testempty``` in the module position field.

Then in your template place this somewhere:
```
<?php
echo "\n\n<pre>==========================\n";
echo 'Module count for position "testempty": ' . $this->countModules('testempty');
echo "\n==========================</pre>\n\n";
?>
```

Before this patch it will show a count of 1, after a count of 0.