Remove readonly on input element, increase size
This change increases the size of the input element by changing the input-small class to input-large, and removes the readonly="readonly" attribute on the input element.

This allows the user to enter in a value into the field, or select a value using the modal as before.

## Testing:

After applying the changes, create or edit an article using the Article Manager, then click on the Images and Links tab. The Intro Image and Full Text Image fields should be larger and a value can be typed/pasted into the field.