A11Y blog categories article count
The Problem: Jaws does not read title tags into the span tag. So the screenreader will only read the number and not the title which discribes the number. Jaws is the most used screenreader in Germany. A lot of blind people reported me this issue.

To create the issue:
-> Use a template without an override for blog_children
-> Make a categorie and subcategories.
-> Put articles into the categories and subcategories
-> create a menu item as categorie blog view
-> Set the following options in the menulink: -> tab Category -> subcategories levels: all
                                                                                        -> #Articles in category: Show
Let JAWS read this and you will not hear the title tag.

Now apply the patch and you will see and hear.
Markus