Removed not useful tests
Unit test should test the interface of a class and not the internals. In this case the getInput Method is tested and this is a protected method. Further more the tests are very basic, they check if properties are set. For the tests reflection and mocking is havily used so it is more a test if reflection works as testing the class. On top the test fails often because now is used and compared later, if this happens at the wrong millsecond the time is different and the test fails.