Remove obsolete/conflicting code
The parent commit attempts to fix a problem with the ENGINE type for table
__update_sites_extension.

The code being removed here was contributing to that problem.

Related history:

Joomla Tracker item 23817 (commit 5db5dd) removed all ENGINE type
specifications from all table definitions.  The intent was to use whatever the
default engine was.  Unfortunately, it seems they overlooked the
__update_sites_extension table in their edits.

Tracker item 26929 (commit 4a2128) added the code that is now being deleted
here.  There's no mention for why the code was added (it's unrelated to the
tracker issue), but the code is obviously attempting to convert the previously
overlooked __update_sites_extension table so that it, too, uses the default
engine, as was the case for all other tables.

Tracker item 28854 (commit 0fc0e9) added the ENGINE=InnoDB specification back
to all tables, including the __update_sites_extension table.

Tracker item 29084 (commit da6182) updates all existing tables (including the
__update_sites_extension table) with ENGINE=InnoDB.  It seems clear that they
were unaware of the code that is now being deleted here, and if they had known
of it, they would have deleted it then because it conflicted with what they
were doing in their commit.