Security Issue with login: password shown in browser
If the user supplies either an unknown user name or a wrong password or a wrong secret key the com_users login form is shown (probably again) and it contains not only the user name typed in but also the password and (if activated) the secret key. The latter two should never be sent back to the user's browser for security reasons as is the case with the registration form when you enter e.g. a user name already in use or an email address already used with any other registration.