Include 'registry' in the migratable keys
The `$_SESSION['__default']['registry']` is responsible for the `userState` as in `$app->setUserState('some.key.for.state', 'value');`
Not migrating this value, as well as not even initializing this value as blank results in `userState` inaccessible until the user log-out and login again. 

This commit includes that key as a migratable key.