Update all instances of hardcoded "administrator" in the entire cms to use a constant for the directory name JADMINISTRATOR_DIR which still defaults to "administrator". Made that constant value configuration from Joomla global configuration.
NOTE: If the administrator directory name is not customised then this would work exactly the same as before.
When using a custom directory name for "administrator" we must ensure none of the installed extension (including 3rd party) are depending on the hardcoded value. Contact the extension developer for support in such cases and not the Joomla community.