Fix hits counter
# Executive summary
These change allows to count hits for an article also when cache is switched on

# Backwards compatibility
Full b/c, no problems are expected

# Translation impact
none

# Why
If you switch cache on hits for an article are not counted right. The counter will be increased for the first time the article is requested. For the time the article come directly from the cache hits are not counted, this makes the statistic useless.

# Testing instruction
## To reproduce the error

* Install a fresh version of Joomla 3.4.8
* Create an article with an intro and full text, we need a read more button
* Make it a featured article so that you can see the article on the home page
* Notice that the counter is 0
* Click on the read more button and go back
* Notice that the counter is 1
* Click again and go back
* Counter === 2

* Switch on cache
* Click on the read more button and go back
* Counter === 3
* Click on the read more button and go back
* Counter === 3 (ups)

You can now wait the cache time and click and the counter will increase. If you are not so patient you can clear the cache and click then and the counter will increase

## Expected behavior
Every time you click on an article to counter should increase if cache is set on or off shouldn’t make a difference.

# Apply patch and testing
* Switch cache off
* check the hit counter value => N
* Click on the read more button and go back
* Counter === N+1
* Switch on cache
* Click on the read more button and go back
* Counter === N+2
* Click on the read more button and go back
* Counter === N+3

Bingo :-)