Remove class="active" from non-menu item
Removed the class="active" from line 17 which is text.  In the standard Joomla! install it says "you are here".  This should not be active, only the menu item for the page you are on should be active.  You can see an example here: http://eoindemo.demojoomla.com/index.php/bread-crumbs-content

Issue posted here: https://github.com/joomla/joomla-cms/issues/8593#issuecomment-172546422

Protostar needs to be updated with this CSS 

.breadcrumb > li {
    color: #999;
}

I am doing this seperately