Correct Styling for Breadcrumb after Pull Request
I have created a pull request https://github.com/joomla/joomla-cms/pull/8929 to remove the class "active" from the "you are here" text and leave it only on the menu item of the page you are on.  The issue is explained better here.  https://github.com/joomla/joomla-cms/issues/8593#issuecomment-172546422

After making this request I have updated the protostar template so the "you are here" text is still the same colour as previous.

The original change may break backwards compatibility with other templates.