Try dc:creator if author is missing in RSS feeds
In feeds from WordPress blogs, the Dublin Core creator tag is used instead of the author tag.  This PR uses the Dublin Core creator tag as a fallback when the author tag is not present in a feed item.

To test, put something like the following code into a template override for the mod_feed module and point the module at a WordPress feed.

foreach ($feed as $item)
{
    echo $item->author->name;
}

You should get a string comprising all the authors of the blog posts in the feed.  You might need to check the raw feed data to see what is actually being sent by WordPress.