Again correct alter table with drop and add
There might be spaces or not before or after the comma separating the substatements, this results in different word indexes.

This code now still assumes that the same index is dropped and added again in the same line, i.e. it does not check if 1st index namd and 2nd index name are the same or not.

And so or so it is not an elegant solution.