Update en-GB page 2 English language
The first sentence is left as a cliff hanger, "one" completes the sentence. 

If the second sentence is still true the randomly generated prefix would be 3 or 4 alpha numerics, to reflect default being 5 I have adjusted the wording to "for or five" characters long, the last part of the sentence is adjusted for better grammer.