Automate utf8mb4 conversion
Automate utf8mb4 conversion for Joomla! core by detecting conversion
statement changes, and extend the conversion status table by column for
extension id to be able to handle extensions, too. Extend fix for
conversion table and add it also to populateState() of the
InstallerModelDatabase class to make sure the table is complete before
the check is run.