Add conversion table fix to script.php
After updating from 3.5.0, the conversion table does not have a record
with extension_id, and it is safer to repair the table in script.php as
doing it with a schema update (where update statements are not checked
by schema manager in case something went wrong. The repari added here is
same as in database.php, except the fix for the complete table missing,
which is not necessary here but necessary there when using the "file
copy and run db fix" update method form any pre-3.5.0.