Update categoryedit.php
When we have a lot of categories to choose, with the same name (tipically in subcategories) it is impossible to choose the right category as a parent category. So i suggest to add the category id at the end of the category name. So we can also search for category id.