JApplicationWeb -setHeader and sendHeader upgrade
Current Joomla header operation causes a duplicate header to replace an original header even when $replace = false.  The replacement is a result of interaction between setHeader and sendHeader.

setHeader happily adds duplicate entries when $replace = false
sendHeader happily replaces headers because its default is to replace

This commit alters both:

setHeader now checks an internal list of headers which may only have one value.
If the header is a single value only, the header array is updated only if $replace = true and the update is ignored otherwise
If the header is not a single value, the header is replaced if $replace = true, otherwise it is added as a multiple value 

sendHeader now creates a new array of values, aggregating values under the header name.  When the header is sent to the header() method, its value is imploded on ', '