Clear cache when user is performing an admin task
I was about to introduce it as a separate plug-in but there is no reason why this could not get into core. This PR make system cache a little bit smarter. Every time user is performing a task (like save, publish, delete, trash etc) plug-in will clear cache so users will get latest page version and administrator will not need to clear it by his own. Login and logout tasks are excluded.