Required password fields moved in preprocessForm
To be able to override form field attribute by plugin, these lines must be moved in the ``preprocessForm`` method.

I wanted to send a custom email to new users. 
If I disable email sending in the joomla user plugin (with ``mail_to_user``), com_users sets ``required`` attribute for password fields to ``true``. It is not what I wanted.