Update index.php
I propose changing the gear icon to a user icon. While I know the user icon is also used in the subsequent drop down, that gear never made sense to me as this is a user menu. There's nothing in that menu that's NOT user related (unless you argue that a log out option isn't user related).

In many other systems, that sort of menu is denoted by a user or head icon. This would just bring a bit more clarity to that menu.

To test the change, log into the admin and see if the cog icon is now a user icon.