Update default_items.php
Added featured-contact to the <li> if the contact is featured.

Please note I used the same code as I used for featured article but I don't know if it works so this needs testing please.  If anyone is a better coder than me please correct my work, and if possible, teach me anything you can.