Allow users to pass in a custom color for background fill
Currently the color to use for non-transparent images' background fill is hardcoded to `black` which provides no flexibility. Allowing the users to define a desired background color fixes this.
This PR adds two new functions `setFillColor` and `getFillColor` and changes the function `resize` to use the defined fillColor, which defaults to `black`.