Fix inconsisteny between new install and update for NOT NULL of urls columns of redirect_links table with MySQL or Microsoft databases (#9759)
* Allow NULL values for redirect_links new_url column.

New schema update sql scripts for mysql and sqlazure to remove the "NOT
NULL" constraint accidently added with PR #9269 for the new_url column
of table redirect_links in the schema update sql script, so it is
consistent with a new installation. For postgresql it seems there was no
change on the "NOT NULL" constraints because they were not changed with
the column changes.

* Rename update sql scripts for 3.6.0

* Handle also postgresql just to be safe

* Correct also the old update sql scripts

Correct also the old update sql scripts for correctly updating pre-3.5.0
versions

* Change back PostgreSQL

For PostgreSQL the PR is not necessary I just saw after testing and
reading manuals.