Correct idx_alias_item_id for mysql utf8mb4
Correct idx_alias_item_id for mysql utf8mb4 and add missing name quotes on other indexes with lengths limits for columns.

The idx_alias_item_id might not have lead to an SQL error before with utf8mb4 in case if not in strict session mode, but it would for sure make problems later.

The missing quotes I have added are cosmetics only and do not do any harm. I texted this script with the modification I proposed here in phpMyAdmin.