new user plugin, simplify user registration
# Executive summary
This new plugin allow registration of a user only with an email address. If the plugin is enabled the username fields is not available in forms. As default also the name field isn’t available but this can be configured.

# Backwards compatibility
A problems can occur when a template overwrite has the changed one of the following views:
* com_users/views/login/default_login.php
* com_users/views/profile/default_core.php
* com_users/views/registration/default.php

Same is possible for the login module.

A small problem can be if someone has a direct link to the „Username Reminder Request“ as we have in our test data. This doesn’t makes sense in this case.

# Translation impact
Two new language files with 8 new language tags.

# Core changes
* New User Plugin Event „onUserBeforeDataValidation“
* New User Plugin Event „onUserAfterDataValidation“
* New JForm Class method getFieldXml as public getter for the protected method findField
* New parameter path for the setField method of JForm, allows to add a field at a specific place in the xml tree.

All core changes are only additions and full B/C, even if the plugin isn’t accepted these changes allow more flexibility for JForm.

# Testing instruction

* Apply the patch
* go to extensions and discover install the new plugin
* enable the plugin
* enable user registration

1) Test registration process
2) Test login
3) Test profile edit frontend
4) Test user data change backend