Make including password in mail to user optional when created by administrator.
When creating user accounts as administrator and sending an email, it is now optional to include the password in the mail.
This is configurable on the plugin. Sending the email was already optional. 
When sending emails, including the password in the email is now optional.

This closes issue https://github.com/joomla/joomla-cms/issues/10376