Enable HTTPs detection for the server scheme "frontend-backend"
The problem appears when using NGINX as the front-end and Apache as a back-end. If the NGINX does encryption, Apache might not see the variable $_SERVER['HTTPS']. In fact, it might be empty, however the URL contains a protocol HTTPs. This commit can help server administrators to handle such configuration. If they set a variable before the back-end, for example `proxy_set_header  X-Forwarded-Proto https;`, Joomla! will see the HTTPs.