Change in file en-GB.com_tags.ini
String changed: COM_TAGS_TAG_VIEW_LIST_DESC
I propose that the sentence "Displays a compact list of items that have been **tagged with** the selected tags" is changed to "Displays a compact list of items that have been **assigned to** the selected tags". The idea here is to avoid the use of the same or similar words while keeping the message clear.