Fix JUri::root() assignment from backend app constuctor (#8587)
* Fix JUri::root() assignment from backend app constuctor

This may not seem any important but I'd say it still is breaking. Try running Joomla in a subdirectory `administrator` under your web root, and open backend. It would break.

* dirname gives a slash for 1st level subdirectory 

Removed trailing slash, which eventually only occurs when queried for a root domain Joomla setup, i.e. not in a subdirectory installation.