More checks needed (Force HTTPS, HTTP(S) ports)
If "Force HTTPS" is set we need to check that the server is available via HTTPS with the HTTPS port probably defined, reset values that might have caused an error. If a HTTP port is set we need to check that the server is available via HTTP with this port, reset value in case of an error. Note that both the HTTP and the HTTPS port are used only if "Force HTTPS" is not set to "None" (0).