Force SSL Check might fail (#11189)
* Force SSL Check might fail

When setting the force SSL option the check might fail if the common name in the server's HTTPS certificate doesn't match the host name used in accessing the server. If it is fine to disable the check whether the certificate was issued by a known CA (by setting CURLOPT_SSL_VERIFYPEER to false) it should be fine too to disable the check for the right common name in the certificate. Both checks might and probably should be delegated to the user.

* Code style change

Blank line before comment added

* Do not use proxy when checking HTTPS

Suggested code change from PR #11193 moved to this PR

* line too long, wrapped

* Code style change