Save an active tab state inside nested tab sets (#10026)
I have modified this code to keep tracks of all active tabs, when there is nested tabs present.
For example, if you making changes in some category and you navigate to the Rules tab there will be another tab set inside with Groups. With previous code only last clicked tab is saved to local storage.
So if you clicked Rules tab in the local storage will be "#rules", but when you click some Group this "#rules" will be overtighten by, for example, "#permission-1" and next time you reload the page your Rules tab will not be active.
With this code all active tabs will be tracked and saved to local storage.