Login form action set to "index.php"
Login form action set to "index.php", in doing so undoing the change made Dec 8, 2014.

Reason for this change is that `JRoute::_()` doesn't honor the third parameter (the "usesecure" flag) when a complete URL is passed, this is due to the following lines in that method:

```
if (!is_array($url) && (strpos($url, '&') !== 0) && (strpos($url, 'index.php') !== 0))
{
        return $url;
}
```

Assume that `http://www.example.org/index.php` is the current URL which is passed as a string as first parameter to `JRoute::_()`. A string is not an array, there is no "&" in the string and the substring "index.php" doesn't start at position 0 in the string.

It is no problem to pass simply "index.php" here as action since the real target after logging in is passed in the hidden field "return".