Plugin-definable fields in login modules / views
This PR contains the PLT-accepted portion of gh-11778

The changes made allow authentication plugins to render their own fields in the login modules and the login views of com_users. This is necessary for social network authentication, single sign on integration and generally any authentication method which does not rely on a Joomla-provided username/password pair.