Solution for tagging in different languages
There is an issue in typing tags in different languages. Pressing ',' in tags means you are done and it will be saved, but it's only for English not other languages such as Arabic, Persian, Hebrew and ... which was bothering me and many other Joomla users. So that I made changes to solve this problem. Now, you can press ',' but it submits tag just if you are typing in English language and if you are typing in other languages continue as long as you want and press enter at the end. Tested in many languages and everything works fine.